Add Transform Equation. Fix Hyperbola quadrant switching.
Hyperbola was not staying in a consistnet plane. Axis was not signed correctly. Signing allows construction across the correct major and minor axes.
Fix rotation of Conic by angle. Input angle now correctly rotates the conic equation.
Renamed major and minor axes to AxisA and AxisB because major and minor are not conisistent due to nature of solver.
Added Conic string format which prints the terms A through B to 8 d.p. into a displayable string.
Renamed TransformShape method to Transform, and renamed Transform property to TransformMatrix for more clarity.