fix parabola 3 point solves domain, parabola computes focus and vertex properly.
Added ParabolaShape which describes the Parabola in a +-x/y domain.
Chunks of the solver have been rewritten to accomodate for Flipped parabola axes. These in general cause the most trouble for the solver.
Perhaps it would be better to move to a unified Parabola solver which generates a normalised parabola in all cases, and then applies axis rotation followed by nominal plane nd equation rotations.